requirements.txt
# LLM Providers
anthropic>=0.39.0
google-generativeai>=0.8.0

# Structured LLM Outputs
instructor>=1.6.0
pydantic>=2.9.0

# Audio
elevenlabs>=1.0.0

# Video Providers
fal-client>=0.5.0
replicate>=1.0.0

# Video Processing
ffmpeg-python>=0.2.0

# CLI & Config
click>=8.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.0.0

# HTTP & Async
httpx>=0.27.0
aiofiles>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# Type checking
mypy>=1.11.0
